Fix some warnings

+++ src/crypto/decryptverifyfilescontroller.cpp
@@ -386,11 +386,11 @@ std::vector< std::shared_ptr<Task> > DecryptVerifyFilesController::Private::buil
     kleo_assert(!useOutDir || outDir.exists());
 
     std::vector<std::shared_ptr<Task> > tasks;
-    for (unsigned int i = 0, end  = fileNames.size(); i != end; ++i)
+    for (int i = 0, end  = fileNames.size(); i != end; ++i)
         try {
             const QDir fileDir = QFileInfo(fileNames[i]).absoluteDir();
             kleo_assert(fileDir.exists());
-            tasks.push_back(taskFromOperationWidget(m_wizard->operationWidget(i), fileNames[i], useOutDir ? outDir : fileDir, overwritePolicy));
+            tasks.push_back(taskFromOperationWidget(m_wizard->operationWidget(static_cast<unsigned int>(i)), fileNames[i], useOutDir ? outDir : fileDir, overwritePolicy));
         } catch (const GpgME::Exception &e) {
             tasks.push_back(Task::makeErrorTask(e.error().code(), QString::fromLocal8Bit(e.what()), fileNames[i]));
         }
